refactor(cli): split update_cmd monolith into a package
Mirror the init_cmd layout: the 1459-line update_cmd.py becomes a package
with focused submodules. Pure code move, no behavior change.

  command.py      Click command + single-repo orchestration
  mode.py         index-only vs full-regen resolution
  incremental.py  thin delegators to core.pipeline.incremental
  persistence.py  index-only persist, ORM->dict, config health re-score
  workspace.py    workspace update flow + editor file refresh
  reporting.py    incremental update report rendering

__init__.py re-exports update_command plus the helpers the test-suite and
sibling modules import, so repowise.cli.commands.update_cmd stays a stable
import surface. Retargets four monkeypatch sites to the submodules that now
own the patched names (run_async, console, _workspace_update, get_head_commit).

Lint posture is unchanged from the original file. All 558 cli unit tests pass.
