+++ tests/unit/cli/test_workspace_autodetect_e2e.py
@@ -58,7 +58,7 @@ def test_update_from_workspace_root_does_not_create_stray_repowise(
         called["agents_md"] = agents_md
 
     monkeypatch.setattr(
-        "repowise.cli.commands.update_cmd._workspace_update",
+        "repowise.cli.commands.update_cmd.command._workspace_update",
         _fake_workspace_update,
     )
 
@@ -82,7 +82,7 @@ def test_update_from_child_repo_stays_single(tmp_path: Path, monkeypatch):
         called["workspace"] = True
 
     monkeypatch.setattr(
-        "repowise.cli.commands.update_cmd._workspace_update",
+        "repowise.cli.commands.update_cmd.command._workspace_update",
         _fake_workspace_update,
     )
 
@@ -91,7 +91,7 @@ def test_update_from_child_repo_stays_single(tmp_path: Path, monkeypatch):
     # the heavy imports by making get_head_commit return the same SHA so
     # the "Already up to date" branch fires.
     monkeypatch.setattr(
-        "repowise.cli.commands.update_cmd.get_head_commit",
+        "repowise.cli.commands.update_cmd.command.get_head_commit",
         lambda _p: "abc123",
     )
